feat(diff): Add support for path types in the DiffFunc signature
This commit modifies the DiffFunc signature in the DiffType struct by adding a new parameter "pathTypes". This parameter is of type []interface{} and represents the types of the elements in the path. The purpose of this change is to provide additional information about the types of elements in the path during the diff process.

Changes:
- Modified the DiffFunc type in the Differ struct to include the "pathTypes" parameter.
- Modified the DiffFunc implementation in the diff.go file to include the "pathTypes" parameter in the function signature.
- Modified the StructValues function in the diff.go file to include the "pathTypes" parameter in the function signature.
- Modified the diff function in the diff.go file to include the "pathTypes" parameter in the function signature.
- Modified the Add function in the Changelog struct to include the "pathTypes" parameter in the function signature.
- Added a new helper function copyAppendType to copy and append elements to a []interface{} slice.
- Modified the diffBool, diffFloat, diffInt, diffInterface, diffMap, and diffPtr functions in their respective files to include the "pathTypes" parameter in the function signature.

The purpose of this change is to provide support for storing the types of elements in the path during the diff process, which can be useful for further analysis and processing.